requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
openpyxl>=3.1.0
rapidfuzz>=3.6.0
sqlalchemy>=2.0.0
plotly>=5.20.0
fpdf2>=2.7.0
reportlab>=4.0.0
xlsxwriter>=3.1.0
pytest>=8.0.0
pyinstaller>=6.0.0
